feat: create TitleCapitalizer
TitleCapitalizer converts an input string to title case.  The
implementation is based on regular expressions (from John Gruber's Title
Case and offshoots) with several additions/changes that I made during
initial development/testing.

For ease of use, it uses the same method naming convention as
Humanizer's `IStringTransformer`.

It provides a nested Builder class is provided to facilitate
parameterized construction, which is desirable due to the many
variations in conventions and style guides, the heuristic nature of the
patterns, and to accommodate application-specific
constraints/needs/heuristics

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>