fix: remove de di name prefix handling
Although de and di are often prefixes of Italian names which cause the
next letter to be capitalized, they also often occur in common English
names where the next letter is not capitalized.

Remove the special handling for now.  Consider re-enabling for Italian
locale (or with much tighter suffix checks).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>