build: Disable SYSLIB1045 warning
Although it would be nice to use `GeneratedRegexAttribute` to generate
`Regex` implementations at compile time, it is only supported when
targeting .NET 7.  Using `#if` for conditional compilation is not
practical due to code duplication (since the attribute argument must be
a literal, the pattern must be duplicated).  Disable SYSLIB1045 until
support for frameworks before .NET 7 is dropped.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>