.editorconfig: Suggest for loop condition (S1994)
Although not checking the value changed by the for_iterator expression
in the for_condition expression of a for loop can be an error, it can
also be intentional behavior, such as using a variable to count retries:

```csharp
for (int attempt = 0; ; attempt++)
{
    try
    {
	...
    }
    catch (Exception)
    when (attempt < 3)
    {
	...
    }
}
```

Although this could be converted to a while statement, it would need an
additional level of indentation to limit the scope of the variable to
the loop.  Doing that, or disabling the warning in every case, seems to
outweigh the value added by this check.  Leave it as a suggestion only.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>